+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Add domain-reload reconnection handling in run Play Mode tests
 - Remove pre-test compilation check from RunTests handler
 
 ## [1.0.0] - 2026-02-26